backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.7.0
pydantic-settings>=2.9.0
slowapi>=0.1.8
python-json-logger>=2.0.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-slugify>=8.0.1
email-validator>=2.0.0

# Database
SQLAlchemy>=2.0.20
alembic>=1.12.0
psycopg2-binary>=2.9.7  # PostgreSQL adapter
SQLAlchemy-Utils>=0.41.1
aiosqlite>=0.19.0  # For async SQLite support
databases[postgresql]>=0.8.0  # For async database access

# API & Web
fastapi-cors>=0.0.6
python-dateutil>=2.8.2
requests>=2.31.0
httpx>=0.25.0

# AI/ML
groq>=0.4.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.1.0
langchain-core>=0.3.0
numpy>=1.26.0
pandas>=2.0.3
scikit-learn>=1.3.0
sentence-transformers>=2.2.2

# Security
bcrypt>=4.0.1

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
pylint>=3.0.2
mypy>=1.5.1
pytest-asyncio>=0.21.1
tenacity>=8.2.3
flake8>=6.1.0
